+++ android/src/main/java/com/pichillilorenzo/flutter_inappbrowser/FlutterWebViewFactory.java
@@ -1,26 +1,20 @@
 package com.pichillilorenzo.flutter_inappbrowser;
 
-import android.app.Activity;
 import android.content.Context;
 
 import java.util.HashMap;
 
-import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.PluginRegistry.Registrar;
 import io.flutter.plugin.common.StandardMessageCodec;
 import io.flutter.plugin.platform.PlatformView;
 import io.flutter.plugin.platform.PlatformViewFactory;
 
 public class FlutterWebViewFactory extends PlatformViewFactory {
-  private final BinaryMessenger messenger;
   private final Registrar registrar;
-  private final Activity activity;
 
-  public FlutterWebViewFactory(Registrar registrar, Activity activity) {
+  public FlutterWebViewFactory(Registrar registrar) {
     super(StandardMessageCodec.INSTANCE);
     this.registrar = registrar;
-    this.messenger = registrar.messenger();
-    this.activity = activity;
   }
 
   @Override